Fix error on FreeBSD
The FreeBSD shm_open() documentation somewhat cryptically says that
"Attempting to open an anonymous shared memory object with O_RDONLY
will fail with EINVAL. All other flags are ignored." As it happens,
the value of the O_RDONLY "flag" is zero, so if you pass no flags it's
the same as passing O_RDONLY.

I don't know if the permission bits argument has any effect for
SHM_ANON objects. Zero permissions work fine when passing an open shm
object to a child process with the same user ID as the parent.