Get pseudo-random data from current time and pid (#2)
Previously we opened /dev/random and read from it which was slower and
more complex. Neither getpid() nor clock_gettime(CLOCK_REALTIME, &tv)
can fail so now we don't need to think about an error return.

With this change we also switch from alphanumeric characters to octal
digits. We don't need as much randomness as the alphanumerics gave us.
With octal and decimal, the ASCII digit range is contiguous so we
don't need an if-statement or char array to look up digits. Compilers
also detect that 8 is a power of two and optimize the division/modulo
digit loop into very simple code (bitwise and/or/shift). Compilers
optimize decimal division too but it's not as simple. Since we have no
particular reason to use decimal, let's use octal.